+++ julia/recursed/rcollatz.jl
@@ -1,29 +1,23 @@
-function stepCount(i::Int128)
+function stepCount(i::UInt128)
     i == 1 && return 0 
     return i = i&1 == 0 ? 1 + stepCount(i >> 1) : 1 + stepCount(3*i + 1) 
 end
 
 
-sArray = Array{Int64}(10)
-nArray = Array{Int64}(10)
+sArray = Array{UInt128}(10)
+nArray = Array{UInt128}(10)
 
-i = Int128(10000)
-j = Int128(10000)
+i = UInt128(10000)
+j = UInt128(10000)
 
 while j > 1
     i = j
-
-    ic = Int128(0)
-#    while i != 1
-#        i = i % 2 == 0 ? i >> 1 : 3i + 1
-#        ic += 1
-#    end
-#    i = j
+    ic = UInt128(0)
     ic = stepCount(i)
     dupF = Int8(0)
     dIndex = Int8(1)
     sIndex = Int8(1)
-    small = Int64(sArray[1])
+    small = UInt128(sArray[1])
     for k in 1:10
         if(sArray[k] < small)
             small = sArray[k]
